Context now delegates matrices into #player#camera, and no longer loads identity before rendering
removing the load identity before render due to a philosophical difference. If the user is expected to manipulate the matrices directly, then load identity would make sense because there are many manipulations to make and the matrix was left in an unknown state at end of render. However, now the user is expected to manipulate the matrix indirectly via the camera, and the camera is expected to "remember" its position between frames. So loading identity in this case would be equivalent to reloading the camera every frame -- which doesn't make sense.