Added a bona-fide matrix stack
The new stack makes all of the operations much easier to understand, not to mention making Jax.Context much cleaner and more efficient.

This commit breaks the current blinn-phong shader. I'm still trying to see if that's because the new matrix stack is broken, or if the shader was built on an already-broken stack.