Implemented a much more robust shader implementation
New impl is more flexible, modular and reusable. It paves the way for attaching multiple shaders to a single material, allowing for better organization and more code reuse. For example, shadow and normal mapping can be removed from the blinn-phong shader and reused for other shaders without duplication of code.

New impl also sets up the Ruby scripts necessary for building custom shaders on a per-app
basis in addition to the built-in ones.