Material and shader chains implemented
Massive changes made to materials and how they interface with shaders. Material/shader chaining now implemented, and shaders are preprocessed to maximize reusability.

Lots of optimizations, etc. still to be made. The textures controller is still completely broken.

Once the fixes are in place, I'll add a detailed wiki entry on how all of this works.