Bugfix: Lighting calculations didn't normalize vNormal
Effect was sort of per-vertex-looking on some models. Effect was the most noticeable on cubes, where specular lighting was greatly exaggerated and would suddenly disappear when the cube rotated away from the light. Using nNormal = normalize(vNormal) resolved the issue.

This issue did not affect directional lighting, which was already using nNormal.