Jax is now smart enough to incrementally back off of shader chains that exceed GPU limitations, instead of throwing a fatal error
So if you have a shader chain [basic, texture, normal map, shadow map, fog] and the fog component is just 1 chain elements too many, the fog option will be dropped. If window.console is available, a warning will be logged; else the drop-off is silent. The shader chain then retries, dropping the last element in the chain until either the chain is empty (nothing left ot drop, fatal error) or until the shader source is within acceptable constraints.

You can control which elements of the chain are dropped by simply reordering them; for instance, in a scene where fog matters more than normal mapping, simply move normal mapping to the end of the chain and fog to the beginning. Jax always tries to drop the last element in the chain first, and works toward the beginning of the chain if necessary.