Automatic tangent and bitangent calculations, and TriangleFan
Tangents are calculated when explicitly requested, or when accessed.
Bitangents too.
Normals are now calculated in a similar manner, so that if they
are not used then they are not calculated.

Tangents and bitangents can be passed into Jax.Mesh.Data, but aren't
yet passed into Jax.Mesh.Base#init to be populated.