Massive performance improvements to Jax.MatrixStack
Translated into CoffeeScript, and simultaneously switched
to lazy evaluation of dependent matrices. For example, when
`#multModelMatrix()` is called, the appropriate ModelView
matrix is _not_ recalculated. Instead, it is merely flagged
as invalid, and recalculated only when it is actually accessed.

In cases where every possible matrix is required by the runtime,
there won't be any performance improvement, but there won't
be any loss either. In other cases where at least one matrix
is not actually used (representing the vast majority of cases),
there will be at least some level of performance improvement.

Benchmarked the process of pushing a level, loading a model
matrix, loading a projection matrix, retrieving the resultant
modelview matrix, and popping the level.

    OLD x 81,710 ops/sec ±3.10% (57 runs sampled)
    NEW x 1,093,916 ops/sec ±0.62% (65 runs sampled)

The results speak for themselves. "Significant improvement"
is an understatement!