Huge performance gain by only rebuilding unique shaders
Previously, each instance of e.g. Jax.Material.Legacy
would build a brand-new shader. Now, shader layers are
added at the class level and only one shader instance
is compiled per material. If anything calls `addLayer`
on the material, the shader layers and source are
cloned, and a new shader is compiled.

Since most materials are just instances of a specific
material class, and most materials don't have layers
dynamically added to them, the result is a massive
performance gain. I haven't benchmarked it, but the
difference is huge.