Use a different shader import name
Apparently use of double underscores ("__") in macro names
is reserved in GLSL, even though it seems to work on most
hardware.

An upcoming release of Chrome has a rewritten ANGLE shader
preprocessor which now enforces this rule, so it's time to
fix Jax -- before the new Chrome goes mainstream.

Firefox will be receiving this update shortly, as well, and
it applies to all operating systems (not just Windows), so
this is a big problem -- but thankfully, an easy fix.