Removed the separate interval for updates
Been waiting for a good reason to do this, and now I have one.
You should be able to (and in the past, were able to) use picking
and other off-screen rendering during controller or model update.
Now, trying to do so introduces a horrible white flickering. This
appears to be related to odd timing in relation to the actual render
process, interrupting it at some level (during back buffer swap?).

The solution is to perform update-related rendering, such as picking,
during the render pass. Hence, update must be performed during the
render pass.

This also removes an internal recurring setTimeout(), so it should
theoretically improve performance and increase battery longevity.