split jax dev manifest into two files so cruft isn't loaded in docs
I noticed this when Firefox spat out a deprecation warning about `java`,
which originated in Benchmark.js. The `require_directory` directive was
pulling in all of the Jax dev tests, bloating the docs js with unused,
and potentially unstable code.

Now the docs require `jax-dependencies.js` and dev mode requires `jax.js`
as usual, which refers to the former to keep it DRY.