Views are now just functions that belong to controllers
By taking the extra layer of instantiating Jax.View around a function
out of the loop, we gain a slight boost to performance, a significant
gain in readability, and less pollution of the parent instance of
Jax.Context. We also no longer have to juggle instance variables between
Jax.Controller and Jax.View. Since the view is just a function and it is
attached to Jax.Controller, it now has direct access to instance variables
and methods of Jax.Controller, including those created by users of the
framework. The user no longer has to do a hacky
`@context.controller.whatever` as they can now do `@whatever` directly.