! API CHANGE - divide by 2 the spot angle given
The spot angle in previous revisions was mathematically accurate, but
produced spot lights with about 2x the cone area as a developer would
expect, because it works in both negative and positive space. E.g. a
spot angle of 30 degrees would produce a visible cone from -30 to +30
degrees, for a "real" cone of 60 degrees. This change causes the angle
to be represented internally as half of the specified angle, but still
returns the specified angle when queried, so that if an angle of 30
degrees is given, the visible cone produced has a -15:+15 = 30 degree
area.

This change will have the side effect of cutting in half the visibel
cones in all current usages.