remove jax binary
We're going to stop supporting standalone Jax apps in v3.0 because
the same can be done with a regular Rails app, with less "magic"
to obfuscate the process.

Having said that, we no longer need the jax binary because in the
context of a Rails app all it does is invoke `rails g jax:...`. Thus,
the long-form `rails g jax:...` is now the preferred invocation for
Jax generators.