stop handling errors implicitly in Jax.Context
Apps should catch errors explicitly when possible, hook into
window.onerror when necessary, and explicitly #startRendering()
and #startUpdating() again when they determine an error is
nonfatal.

This was a failed attempt to generalize error handling where
such generalization was neither a necessary nor core responsibility
of Jax. Further, the implementation masks errors entirely in some
recent browsers unless an error handler was explicitly added by
the dev, which is an unexpected and undesired side effect.